requirements.txt
streamlit==1.33.0
streamlit-drawable-canvas==0.8.0
Pillow
pandas